move addStakeAmountChange
